Add option to specify date format
This must currently be specified by setting a constant's value.
In the future, I'll probably try to add an input field for this.
The default date format is set in a way that the regex is original.
Other placeholders for date items might be required in the future,
but that should be a trivial task.