Add information about the possibility to set localization strings
I put it into footer, because there was already a style prepared for it.
I originally wanted to put an input field for those two pieces
of information somewhere inside the HTML, but I just couldn'd find
a fitting place and not that many people will probably need
to touch the original values.